(CPR) 90% done, almost ready to send to fab
Power Strip board is 100% complete.
Controller board is 95% complete, all that's left is to figure out a method to restart into Boot Select with a single button press, and double-check that it works before sending it off to print.

Remaining space of board is being filled out, and is nearly finished.

I'm considering adding plated holes for spacer holes into the Power Strip board.  We'll see.